+++ performance_test/CHANGELOG.md
@@ -7,6 +7,7 @@
   perf_test will use `/dev/cpu_dma_latency` to request that the CPU not enter
   any sleep states, to potentially give more consistent results
 - Some smaller `Array` messages, down to 32 bits
+- Added support to the FastDDS plugin for bounded and unbounded sequences
 ### Changed
 - Update the README to better explain how to use this tool with Apex.OS
 - In the `Runner`, allocate the `AnalysisResult`s on the stack instead of using `shared_ptr`